Elaborate on the project's goals in the README
In addition to explaining the goals and usage of this project, address
the `bundler`-generated `atomic.gemspec` warnings:

```
You have one or more invalid gemspecs that need to be fixed.
The gemspec at .../atomic/atomic.gemspec is not valid. Please fix this gemspec.
The validation error was '"FIXME" or "TODO" is not a description'
```