Support fallback to the underlying helper method
The `Atomic::Fallback` wraps the decorated Rails helper in a way that
can be invoked at a later point in time.

---

In some cases, the changes in the Atomic view partial don't amount to
much more than default `class` declarations.

There are two ways to leverage the overridden, Rails-provided helper:

1. invoke the helper directly, by name, and splat out the view partial-local
   `arguments`, `options`, and `&block` variables:

   ```html+erb
   <%# app/views/atomic/tags/_h1.html.erb %>
   <%= tag.h1(*arguments, **options.with_token_list_defaults(class: "heading"), &block) %>
   ```

2. chain your calls to declare default token lists (via
   `with_token_list_defaults`) or other default attributes (via `with_defaults`)
   off a reference to the partial-local `fallback` variable:

   ```html+erb
   <%# app/views/atomic/tags/_h1.html.erb %>
   <%= fallback.with_token_list_defaults(class: "heading") %>
   ```